Refactor AbstractTestRest to support paralell hierarchy
