+++ servers/jax-rs-tests/src/test/java/org/projectnessie/jaxrs/TestJerseyRestNaiveClientInMemory.java
@@ -18,6 +18,7 @@ package org.projectnessie.jaxrs;
 import static org.assertj.core.api.Assumptions.assumeThat;
 import static org.projectnessie.client.http.HttpUtils.HEADER_ACCEPT;
 
+import java.net.URI;
 import org.jetbrains.annotations.Nullable;
 import org.projectnessie.client.api.NessieApiV1;
 import org.projectnessie.client.http.HttpAuthentication;
@@ -48,7 +49,7 @@ import org.projectnessie.versioned.persist.tests.extension.NessieExternalDatabas
 class TestJerseyRestNaiveClientInMemory extends AbstractTestJerseyRest {
 
   @Override
-  protected void init(NessieApiV1 api, @Nullable HttpClient.Builder httpClient) {
+  protected void init(NessieApiV1 api, @Nullable HttpClient.Builder httpClient, URI uri) {
     assumeThat(httpClient).isNotNull();
 
     // Intentionally remove the `Accept` header from requests.
@@ -67,6 +68,6 @@ class TestJerseyRestNaiveClientInMemory extends AbstractTestJerseyRest {
             .withUri(httpClient.getBaseUri())
             .build(NessieApiV1.class);
 
-    super.init(api, httpClient);
+    super.init(api, httpClient, uri);
   }
 }
